Expose instanceName in the subject template
Add a variable `instanceName` to the variables offered in the subject
template.

Remove the instanceAndProjectName as it is easy to reconstruct it using
the projectName and instanceName variables. It was also incorrect, as
it relied on the short name and not on the project name.

It was introduced recently, so _I_ don't think any gerrit instance will
break.

Change-Id: Iaf46fcbc1df6db637cb95711bac84a4ead5fc036

Former-commit-id: eb4d2d8da7bfc29511154fffe19f89f769a6a03c