Prevent duplicate permissions in AccessSection
Duplicate permissions in an access section are not allowed. This is why
AccessSection#setPermissions(List<Permission>) fails with an
IllegalArgumentException if the provided list contains duplicate
permissions. However by calling AccessSection#addPermission(Permission)
it was possible to add duplicate permissions.

Duplicate permissions on an access section of the All-Projects project
make the permissions of the All-Project project unreadable and Gerrit
effectively stops working. This is because
ProjectState#getLocalAccessSections() filters out some permissions on
the All-Projects project by getting the list of permissions from
AccessSection and setting the filtered list back on AccessSection. If
the list that was retrieved from AccessSection contained duplicate
permissions, setting back the list failed with IllegalArgumentException
since AccessSection#setPermissions(List<Permission>) doesn't allow
duplicate permissions.

Also see change I29298232a3 which demonstrates this bug.

This change makes the behavior of the methods in AccessSection
consistent, AccessSection#setPermissions(List<Permission>) and
AccessSection#addPermission(Permission) now both reject duplicate
permissions.

The caller of AccessSection#addPermission(Permission) that tried to add
duplicate permissions was removed by change I139ba7aa97 which reverted
the in-place migration of refs/for/ permissions. When this code is added
back the new implementation must care about merging duplicate
permissions, otherwise the migration code will now fail.

In general it's better to fail at this earlier stage and prevent the
creation of a AccessSection with duplicate permissions as this prevents
that such a bad AccessSection is ever stored.

Change-Id: I5222cd9174f0b17baae2e66702c4aec6133a14ba
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 8044bb1b8789b1f3b4bbc695e96bdb2ec179f677