Remove documentation about cleaning the build cache
The buck-cache folder is not used since change Ia9c58bc5a, and the whole
process of cleaning the build cache is different with bazel anyway.

Rather than attempting to document this for bazel, where we rarely need
to wipe the build cache (compared to with buck), just remove that part of
the documentation.

Change-Id: I1ae4e89f1b9c9a8cc3ce75a18a6e48db2d1b2ba2

Former-commit-id: 89050c03bc2eeece0f4384ab97e02cd441705fe6