change.allowBlame config only disables in the GWT UI
Because of problems in the GWT UI's display of blame annotations, the
change.allowBlame config was used to disable the Gerrit feature
altogether. Since then, the PolyGerrit UI can display blame annotations
without these problems, but the feature remains disabled.

The change.allowBlame config handling is changed to only disable blame
in the frontend of the GWT UI. Note: this removes the ability to disable
the blame API via server config.

Change-Id: Ic25aed5a533e0044956359587321676b022c21ec

Former-commit-id: 78ef01e3afa7b4040cc442154a3dc4b37e661004