Convert OAuthTokenCache serialization to protobuf
The serializer class lives in OAuthTokenCache.java so the extension
framework doesn't acquire a dep on the cache serialization code, which
currently lives in the server package.

Proto3 doesn't have optional fields, so we can't distinguish between
empty and null for the provider_id field (at least without adding a
separate has_provider_id field). Explicitly treat these the same in
OAuthToken itself, and clarify that other fields are not nullable.

Change-Id: I37bd47ac0c8491e5212fcec4885e063ae50d714d

Former-commit-id: e6930d73b01d68f265a78dbbb3ecc449b0b77f55