Add tests for fields declared in types with custom serializers
When the set of fields changed, it's a very good sign that something
needs to change in the serializer implementation, or at least bumping
the version number in the CacheMoule. Add a simple test utility that
maps declared field names to types, and list the expected types
explicitly in each of the relevant tests.

This simple mechanism can't catch everything, for example:

 * If the serializer implementation changes for an existing type in an
   incompatible way, without altering any fields. In this case, the
   {key,value}Serializer in the CacheModule will change, so hopefully a
   reviewer will notice and ask the author to bump the version as well.
 * If semantics of individual fields change without changing the type,
   such as making a field newly @Nullable.

But it's much better than nothing.

Change-Id: I51e60b83f9c3291d2f8f9ac19ce0192678211676

Former-commit-id: 095394c06995a8e3236cee65f4583b19b85bd631