Fix the online index for changes when upgrading ES gerrit site
Previously, when using ElasticSearch as main index engine and
upgrading the gerrit site from 2.14 to 2.15, the changes were not
visible in the UI even when OnlineIndexer logs showed that indexing
was successful.

Reason: After migration, new indices were never created with the custom
mappings and elasticsearch created the default mappings with the
standard fields used in the indices. This resulted in a mismatch between
the mapping types and therefore changes were not visible in the UI.

For example:

The "changes" index in gerrit use two types of mappings - "open_changes"
and "closed_changes". When no index was created by gerrit and
onlineIndexer was called while running gerrit with the migrated site,
elasticsearch created only "open_changes" mapping type resulting in
conflicts between two different versions of the same indices.

This change resolves the inconsistencies between offline and online
reindex by deleting all the previously existing indices and creating the
new ones.

Bug: Issue 8806
Change-Id: I51ee5bfeb2a154630955b3c94b78fda7d050610d

Former-commit-id: 192dfd7bc6dc9b1d363180d1fbacff743f2412c9