PostReview: Use correct patch set ID to reject comment on non-existing file
When comments are posted for a revision it is checked that the files for
which the comments should be added exist in the revision. If not the
request is rejected with 400 Bad Request. In the error message we should
use the patch set ID of the revision on which the comments are posted,
but we always used the patch set ID of the current revision.

Change-Id: I95a505d95fdf4ee104963adb9f935c17f704f868
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: a284feacadd930111198971d458361fd97493102