Add a helper method to fetch the commit message
The PRED_commit_message_1 rule doesn't need the whole PatchSetInfo
values to be loaded to function properly: only the commit message is
required.

This change improves this by only loading the commit message from the
repository.
Fix a test case: the commit_message predicate does not use the account
index anymore. Before that, it was indirectly used to gather
informations on the author and committer. These two informations were
loaded but never used.

Change-Id: I3a964bae3512ee6fba7895739f32fd592e8595bf

Former-commit-id: 05ce4c24147d162f5768b83bc0733d823a245aaf