Update gr-thread-list to less eagerly re-filter
Previously, showing and hiding drafts/unresolved in the thread list was
done via CSS, based on a reflected attribute of the thread list.
Because the attribute updated immediately, the CSS effect took place
when unwanted.

Instead of using CSS to hide and show items, this change filters them
instead.

This change also makes one fundamental change in the filtering of
threads-- threads are considered unresolved/resolved based on the latest
non-draft comment.  If a thread consists of only a single draft comment,
the unresolved bit is taken from that comment.

Furthermore, this change re-sorts and filters threads whenever a toggle
is selected, but not when a comment thread is modified on its own.

Bug: Issue 8713
Change-Id: I6f4457dbd193e3b2c60caff1e1178972dae9207c

Former-commit-id: 14eb78ff3a6eab76381b67e0b2c2da579df7fc39