Toggle diff expand on keyup
If a user holds down the [i] or [Shift]+[i] keyboard shortcut keys to
toggle inline diff expanded state, they may generate a large number of
toggle events continuously until the shortcut keys are released. This
was because the key bindings listened for keypresse events.

With this change, the bindings listen to keyup events, which are not
emitted continuously when held down.

Bug: Issue 8115
Change-Id: I803a48f9912e6dead7ec439a7694d2412265afa5

Former-commit-id: f4eed8754fc1b9839deb5db5632483fbbb812b97