ChangeQueryBuilder: Allow ownerin predicate to be evaluated by the index
The previous implementation of ownerin predicate was not efficient as
it first found all the matched changes and then checked the membership
of the owner for each one of them. This change modifies this predicate
such that for each account in the group, it finds all the matched
changes that belong to that account, which can be evaluated by the
index.

Note that for external groups (e.g., ldap groups), it falls back to the
previous implementation of this ownerin predicate, as the account lists
of external groups are not available in the index.

Bug: Issue 8916
Change-Id: I57b0a33c824f103d8b0e5e533af30e7cb23aa41d

Former-commit-id: 65a1ae50eb8af57b1e29e8bda34d25f736968869