Disable draft magic branch option per default
repo-tool still doesn't support private and work-in-progress change
workflow. This is the reason, draft magic branch option is still
supported. For other projects, that do not depend on repo-tool it's
very confusing that the draft push option is still supported, even
though the draft workflow was replaced by work-in-progress workflow.

There were attempts to alleviate the situation, with a number of
different approaches, like: I8a992d8d7b and Ie585bb48cc, but the
review was staled.

Bug: Issue 8933
Change-Id: Ia15ffbb430a1de970ca778a36b6fa50021c9b19d

Former-commit-id: d1cb918d3b11f6e32f32d9994e6cbe92cb604d7d