Propagate PermissionBackendException in DefaultRefFilter
Now that RefFilter is part of PermissionBackend and the interface
definition throws PermissionBackendException we can propagate this
exception from when it is throw.

This has the risk of breaking ref-advertisement in cases where we
previously just dropped the exception, but in general makes the code
easier to maintain as we see issues faster.

This commit also removes the duplicate instantiation of
PermissionBackend#ForProject.

Change-Id: If84ed264779b0729dc8c07e589cebade9070f883

Former-commit-id: 2d79d9b6fab42a0b5953c27f3c36f0f91fd1ee5f