Improve error message for submit rules in Submit action
Previously, we would display "This change depends on other changes
which are not ready" when encountering a problem with the submit rules
in the Submit action.

This commit uses the exception's message instead. Since
ResourceConflictException is a REST API exception, displaying it's
message to a user is allowed.

Also improve the other error messages to include the change ID, when
doing so is a reasonable action (the change is not hidden).

Change-Id: I3b050c65844f3cdc21285b8477fec0a8bf123db5

Former-commit-id: d1cefeb300767941db952e03d44e04218a2db8ca