WorkQueue: Don't lazy-initialize the default queue
The default queue is used in several places and is initialized during
the server start-up, so the lazy initialization doesn't win much.

Moving the initialization into the constructor means we can mark the
queue as final, and don't need to keep the default queue size value as
a member.

The getDefaultQueue no longer needs to be synchronized, since there is
no longer any chance of racing to create the queue.

Change-Id: Ib880e3b9caf79718865944ec88c3337d0080e2c3

Former-commit-id: 9c1a6a012eeb655a16d5c932bcf98c2aa2d264d1