Migrate restapi classes to Flogger
This is the next part of the migration to Flogger. This change migrates
all classes of the 'restapi' module to Flogger. Classes of the 'httpd'
and 'server' modules have been migrated by predecessor changes. Other
modules continue to use slf4j. They should be migrated by follow-up
changes.

During this migration we try to make the log statements more consistent:
- avoid string concatenation
- avoid usage of String.format(...)

Change-Id: I11ef03118e5c104bd2e11cb0ad9d37e09514355f
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 6dab32b4ddd93b7d0635c49b9e429b630766d385