Add dependency to Flogger
Flogger is an open source project (Apache 2 License) from Google that
provides a fluent JAVA API for logging:
* https://github.com/google/flogger
* https://google.github.io/flogger/

We want to use Flogger instead of slf4j. This idea was discussed at the
Gerrit hackathon in Lund [1] and on the mailing list [2].

Benefits:
* Better Readability
* Better Performance
* Will allow us to implement request tracing

Flogger API examples:
* logger.atSevere()
    .withCause(err)
    .log("Error in %s %s", req.getMethod(), uri);
* logger.atInfo()
    .atMostEvery(1, TimeUnit.HOURS)
    .log(...)
* logger.atInfo()
    .withStackTrace(StackSize.FULL)
    .log(...)

Within Google Flogger is long-established and is used by almost 3/4 of
all Google's Java projects, hence this library is very mature.

The Flogger API can be used with different logging backends. We are
using the log4j backend, as we are currently using log4j as backend with
slf4j. This means there are no changes to how logs are configured
(existing log4j.properties and log4j.xml files continue to work). It
also means that this doesn't effect any integrations that work with
log4j (e.g. integration with logstash continues to work).

To use the log4j backend for Flogger the system property
'flogger.backend_factory' must be set to
'com.google.common.flogger.backend.log4j.Log4jBackendFactory#getInstance'.
E.g. this system property may be set in gerrit.config:
[container]
  javaOptions = "-Dflogger.backend_factory=com.google.common.flogger.backend.log4j.Log4jBackendFactory#getInstance"

If this property is not set Gerrit sets it automatically on startup so
that we always have a proper logging configuration. We do this in
Main.java before loading any class. This is important since Flogger
reads this system property only once when the FluentLogger class is
loaded and we must be sure that the system property was set before this.

Using Flogger and slf4j in parallel is not a problem. Logs from both
APIs end up in the same log file(s).

This change doesn't migrate any classes to Flogger yet. This will be
done by follow-up changes.

JGit is still using slf4j and needs an explicit runtime dependency on
slf4j.

[1] https://docs.google.com/presentation/d/1Kwn77c1PScMLDEWiJDklEm43NsN2-mWm2PqHwYTkJfI/edit?usp=sharing
[2] https://groups.google.com/d/msg/repo-discuss/meEWJO7c4pA/fxDx2waRCQAJ

Change-Id: Ifc31fb95c10397857e8de5657c729f49c1820ec4
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 2b81d71374fd61937f41a02ce16dbf6484ff3006