Instrument RestView#getDescription to analyze slow UIActions
UIActions (GetRevisionActions) calls have a 99.9%ile latency > 1 minute.
We only have metrics on the REST endpoint which does not give much
indication on why this endpoint is slow.

UIActions is implemented in a way that it calls #getDescription on all
RestViews that implement the UIActions interface. This means we do
serial calls to a lot of these methods. This commit adds a latency
metric to instrument these calls. Knowing which actions are slow can
guide futher steps that we take.

Possibilities include reworking slow actions or looking into more
parallelism.

Change-Id: I261bfb1a5a43b81f11dba019d405126198908c5f

Former-commit-id: 4a56ff01710a08324cd81d88fa4879b7f66ccac6