Report average timings
The gr-reporting tool exports a pair of methods for timing arbitrary
interactions: `time(name)` and `timeEnd(name)`. However, many of these
timings span repeated operations such as rendering some number of file
rows. Consequently, when viewed in aggregate, the absolute time taken to
do all the repeated operations provides incomplete context to identify
performance degradations.

Add a new method to gr-reporting as an alternative to `timeEnd` named
`timeEndWithAverage`. This method is used in the same way as `timeEnd`,
but takes an additional `denominator` argument used to divide the
overall time and find an average. It also takes an additional label
that's used in reporting the average value.

Note: the average is reported in addition to the absolute time. While
`timeEnd` reports only the absolute time, `timeEndWithAverage` reports
the absolute and the average separately.

Update gr-file-list to report the average time to render however many
files were included in the timed batch.

Change-Id: I561078f6b0b2aeed876a86de3828f09b604bd959

Former-commit-id: 94d1302845920cdc67503ec87acaa8cbb1faf438