Allow persistent caches to be not persisted by default
H2CacheFactory already respected a non-positive value and disabled
persistence when explicitly configured in gerrit.config. However, the
implementation of PersistentCacheProvider would never return a
non-positive value from diskLimit(), so it was impossible to disable
persistence by default. Fix this so PersistentCacheProvider uses 128 MiB
when the instance is created, but allow overriding with a non-positive
value.

Change-Id: I189f6dad2003c8bb8ba6e05705e2903c099759c3

Former-commit-id: dab51552590c8493b215a42909a540d1e1601cf9