Add GerritRuntime enum to describe the current running environment
Gerrit can run in different conditions: as a Daemon server, in batch
mode... Gerrit's behaviour should be adjusted accordingly, and with this
new enum it is possible to do it.
It is bound in the configuration injector, so the sys modules can depend
on it to enable (or not) different modules.

Change-Id: Id811f1447a8078afccc96bbfbc707a761aa103a1

Former-commit-id: 020977e53bf3428a34c864f399564d0bd3477c81