Enable plugins in BATCH mode
Batch mode is defined (for now) as "everything that is not Gerrit
Daemon". This includes the "reindex" task, but is not limited to it.

Adding plugins to batch mode allows plugins with a direct impact on
the changes' index to have a consistent impact: without it, custom
prolog predicates and plugin-provided SubmitRule-s are ignored.

In BATCH mode, Gerrit will try to find and load Gerrit-BatchModule(s)
modules defined by plugins, if any. This module should be compatible
with the batch mode injectors: it must not rely on SSH or HTTP classes.
This new module is optional, and plugins not defining one will not be
loaded in batch mode. As such, this new feature is retro-compatible: no
"old" plugins will be loaded.

Change-Id: I9be5a2ea0fd841986b940e046d215a820497107b

Former-commit-id: 5ccf0b9e39afc9372d9867c4337e81751bbcbd8b