Align ChangeColumns methods with Change fields
The documentation previously specified that method order in
ChangeColumns should match @Column order in Change, but this was not
actually true. Fix it.

Rename some ChangeColumns fields to match the underlying column names in
Change. In some cases, there is a mismatch in Change between the field
name and the method name, for example reviewStarted vs.
hasReviewStarted(). ChangeColumns is an AutoValue where the method names
are analogous to fields, so prefer the field name when there is a
mismatch.

Change-Id: I8ba34146bc23832c42f52e5b05bf99d194b1060a

Former-commit-id: 1a647647f8db448c5c7a28802a86e3906351525a