Refactor optional _getChangeURLAndFetch arguments to objects
The _getChangeURLAndFetch method took many optional arguments, making
call sites difficult to read and making the method difficult to make
changes to.

Refactor the optional arguments into an object with optional properties.
The new Defs.ChangeFetchRequest ensures that the objects are well-typed
and call sites are updated as well as tests.

Feature: Issue 8324
Change-Id: Icc1265613d84008a794670d3730136b1ade047f5

Former-commit-id: 366cd4fe7a99401be1c3f7a4108c51610a5d5190