Remove ValidationError#createLoggerSink to avoid passing around loggers
Passing loggers between classes is bad and we should avoid this if
possible.

ValidationError#createLoggerSink required to pass in a logger just to
return it back to the caller wrapped in a lambda. Remove this method and
instead create the lambda directly in the caller. This brings us minimal
code duplication for formatting the message, but it's better than
passing the logger around.

Change-Id: I6b475de50fb2a1745c16ac424e0a95d358aaaf7d
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: ba6fd0f3b117108057a7bdc3288704f8ee1c04bd