WorkQueue: Don't fail when queue metric already exists
It is possible to create mutiple queues with the same name, however the
metric maker implementation (in this case dropwizard, but it may be the
same case for other implementations) disallows to create a metric if
one with the given name already exist.

A specific case where a queue with the same name may be created is when
a plugin creates a queue. When the plugin is reloaded, a queue with the
same name gets created.

Handle the exception raised by the metric maker and if it was due to the
queue already existing, log a warning rather than failing.

Change-Id: I74a0c8b8363ba3774b1cad0b0c27ad437271c81d

Former-commit-id: 4e97e35b7c3423ff4a85a02d4b03722f903d6348