Fix some Javadoc links which are reported to be broken
Our Bazel build issued three warnings about broken Javadoc links. Inside
of an IDE, those links can be resolved without issues, though.

The problematic part seems to be the use of inner classes as method
parameters in the link definition. The warning goes away if we add the
enclosing class.

Bug: Issue 8845
Change-Id: Ic0332c50852680d05880857f0fe285ae5954503c

Former-commit-id: ad0caab5030c2dcbb2a791992453ecc1e63a243d