Deduplicate conditions on current change in Submit
Before this commit, the change was part of the change set on which the
permissions and submit records were checked. As such, these checks were
done twice, using potentially invalid cached submit records (stale).

Instead, we want to ignore the change and only consider the parent (and
the topic) changes.

Change-Id: I28c302db7f8f153903493d999eb587527681d895

Former-commit-id: 62d9e3637d1226f6c8f8d3b69d00d9c04be79b64