Remove unused opt_ctx param from API interface
The `opt_ctx` parameter allowed the callers of REST API interface
methods to specify the `this` context in which the `opt_onErr` handler
would be called. This provision was a vestige of the time before arrow
functions made it easy to have callbacks executed in the context of
their defining element, and was almost completely unused by this point.

Remove the optional param and update callers. The one caller that used
the parameter didn't really need it because the `opt_onError` handler
had already been rewritten as an arrow function.

Bug: Issue 8324
Change-Id: Ifc4d4fbb29561ce140055b86550a8ea9025cb6eb

Former-commit-id: a7474e22e02fd36c895eed1ae5be0815187f710f