Acceptance tests: Remove ES bootstrap boilerplate
Instead of initializing the indices and mapping using Transport client,
just use the production code for that: AbstractElasticIndex.deleteAll()
does the job. Refactor the code to access the injector and call this
method iterating over Collection<IndexDefinition<?, ?, ?>>. That way
we also don't need clean up anymore, because deleteAll does two things
delete all indices and create the mapping.

With this change embedded ES mode is still used for acceptance tests,
but only Node API to start a single node of cluster.

Change-Id: I57e5d0718424b8ff1593f5152b5b1113bd0b6a8e

Former-commit-id: 764af7d0b5b95a664dda4732eb6702ddd769d0ee