Provide the static resource path via index
When the side loader for the syntax library was written, it needed to be
compatible with CDN configurations, but at that time the simplest way to
determine the CDN's path was by extracting it from the URL of the
application's main link element.

Since that time, the CDN path is provided to the page's index Soy
template using a trusted string. This can be safely inserted into the
page's JS block of constants and provided to the loader more directly.

Change-Id: I3a3432c1bf99a650645a7adf1d6979a900988ac4

Former-commit-id: e60796ac6f53e9af4ed1d44c5614194c4e36c69f