ChangeKindCacheImpl: Don't use reference equality for trees
It is basically an implementation detail of the JGit RevWalk machinery
that the getTree() methods for two RevCommits from the same RevWalk
return the same RevTree instance. Although it may be a documented and
stable detail, the code is easier to reason about if we don't depend on
it. This particular code is in a cache loader, where we are certainly
not sensitive to the performance difference between == and
Objects#equals.

Change-Id: I7ae66ac80c1344560543692f6631f469f038c979

Former-commit-id: 31a2e952602981393cfd14bc9acadf7ab25fb298