Ensure files array is not computed early
I9a166dfb introduced a bug in which the file list display state could be
computed before the request for reviewed files was completed.

This change introduces an additional check for the _files array
computation that ensures loading all data has completed. If the file
data is still loading, computation of the _files array terminates early.

Bug: Issue 9057
Change-Id: If51d55a03b7974744b98bfc22967a09ac05e9086

Former-commit-id: b26606c2f2489496e875f3aaddea41f217806dcf