Log timing for all REST calls
All REST calls that go through the REST API interface log their results
to the JS console. Log entries include the HTTP status of the response,
the HTTP method used, the elapsed time in milliseconds and the original
URL of the request.

Note: because the URLs are not anonymized/sanitized these logs are not
sent to gr-reporting.

Change-Id: Ie9c50c91641fb7e63b5e4fb3403cc0e84aca8c8e

Former-commit-id: 2c0bc302d581a0fc0949fcf394b5b56238a3e942