Add 'docker' tag to build rules for Elasticsearch tests
Split ElasticReindexIT out to its own rule with the labels 'docker' and
'elastic'.

Add a 'docker' label on the Elasticsearch query tests.

This will allow to exclude all the tests that require docker:

  bazel test --test_tag_filters=-docker //...

Currently only the Elasticsearch related tests use docker, so the same
could be achieved by excluding the 'elastic' tag, however it is possible
that in future we will add more tests that use different docker
containers.

Change-Id: Ia67835d96c5863a14a060505f1aece7f4b39a12b

Former-commit-id: 7e18f1a84a51a247bee7b51ea227a1e4fcf285e6