AbstractElasticIndex: Move generation of index name to ElasticConfiguration
ElasticConfiguration already reads all of the other configurations, so
move reading of the prefix there too.

Move generation of the indexName from AbstractElasticIndex to a helper
method in ElasticConfiguration.

AbstractElasticIndex no longer needs the @GerritServerConfig annotated
Config, and now just gets ElasticConfiguration injected.

Modify ElasticConfiguration to keep a reference to the @GerritServerConfig
annotated Config, and provide an accessor method.

ElasticVersionManager also no longer needs to get the @GerritServerConfig
annotated. It just gets the ElasticConfiguration which it uses to get
the prefix, and passes the underlying Config up to the superclass which
uses it to get the onlineReindex flag.

Change-Id: Ie4696b4d5a4e0bc58d3b2175f46002d0cc65940c

Former-commit-id: 390c042ab4deb35304aea69a0e64911e94e46807