ElasticVersionManager: Use correct configuration value for prefix
In change I4c6e68783 the configuration option was renamed from
index.prefix to elasticsearch.prefix, and since change I9d089b6a0 it
is not set by default.

As a result, the ElasticIndexVersionDiscovery.discover does not get the
expected results from the _aliases query, and the available versions
are not returned.

Bug: Issue 9041
Change-Id: Ibd00827e5462d6d47cc8a250ba6d29232dccefea

Former-commit-id: 63e278fe4439f381190470ce5e6601c18c23c23f